Did reorg in python locations
