+++ internal/__init__.py
@@ -14,18 +14,3 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from unittest import mock, TestCase
-from boto3 import Session
-from ...src.app.remediations.aws_ec2_terminate_instance import AwsEc2TerminateInstance
-
-
-class TestAwsEc2TerminateInstance(TestCase):
-
-    @mock.patch.object(Session, 'client')
-    def test_fix(self, mock_session: mock.MagicMock) -> None:
-        mock_client = mock.Mock()
-        mock_session.return_value = mock_client
-        resource = {'Id': 'TestInstanceId'}
-        AwsEc2TerminateInstance()._fix(Session, resource, {})
-        mock_session.assert_called_once_with('ec2')
-        mock_client.terminate_instances.assert_called_once_with(InstanceIds=['TestInstanceId'])
