+++ internal/compliance/remediation_aws/src/app/remediations/aws_cloudtrail_enable_log_validation.py
@@ -18,22 +18,22 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsGuardDutyCreateDetector(RemediationBase):
-    """Remediation that creates a GuardDuty detector if one doesn't exist"""
+class AwsCloudTrailEnableLogValidation(RemediationBase):
+    """Remediation that enables log validation for existing CloudTrail trail"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'GuardDuty.CreateDetector'
+        return 'CloudTrail.EnableLogValidation'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
-        return {'FindingPublishingFrequency': 'FIFTEEN_MINUTES'}
+        return {}
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        session.client("guardduty").create_detector(Enable=True, FindingPublishingFrequency=parameters['FindingPublishingFrequency'])
+        session.client('cloudtrail').update_trail(Name=resource['Name'], EnableLogFileValidation=True)
