+++ internal/compliance/remediation_aws/src/app/remediations/aws_ddb_encrypt_table.py
@@ -18,30 +18,22 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsS3BlockBucketPublicAccess(RemediationBase):
-    """Remediation that puts an S3 bucket block public access configuration"""
+class AwsDdbEncryptTable(RemediationBase):
+    """Remediation that creates a KMS key and uses it to encrypt DDB table"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'S3.BlockBucketPublicAccess'
+        return 'DDB.EncryptTable'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
-        return {'BlockPublicAcls': 'true', 'IgnorePublicAcls': 'true', 'BlockPublicPolicy': 'true', 'RestrictPublicBuckets': 'true'}
+        return {}
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        session.client('s3').put_public_access_block(
-            Bucket=resource['Name'],
-            PublicAccessBlockConfiguration={
-                'BlockPublicAcls': parameters['BlockPublicAcls'].lower() == 'true',
-                'IgnorePublicAcls': parameters['IgnorePublicAcls'].lower() == 'true',
-                'BlockPublicPolicy': parameters['BlockPublicPolicy'].lower() == 'true',
-                'RestrictPublicBuckets': parameters['RestrictPublicBuckets'].lower() == 'true',
-            },
-        )
+        session.client('dynamodb').update_table(TableName=resource['Name'], SSESpecification={'Enabled': True, 'SSEType': 'KMS'})
