+++ internal/compliance/remediation_aws/src/app/remediations/aws_ec2_stop_instance.py
@@ -18,17 +18,17 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsRdsDisableInstancePublicAccess(RemediationBase):
-    """Remediation that disables public access for an RDS instance"""
+class AwsEc2StopInstance(RemediationBase):
+    """Remediation that stops an EC2 instance"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'RDS.DisableInstancePublicAccess'
+        return 'EC2.StopInstance'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
@@ -36,4 +36,6 @@ class AwsRdsDisableInstancePublicAccess(RemediationBase):
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        session.client('rds').modify_db_instance(DBInstanceIdentifier=resource['Id'], PubliclyAccessible=False)
+        session.client('ec2').stop_instances(InstanceIds=[
+            resource['Id'],
+        ])
