+++ internal/compliance/remediation_aws/src/app/remediations/aws_kms_enable_key_rotation.py
@@ -18,8 +18,8 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
