+++ internal/compliance/remediation_aws/src/app/remediations/aws_rds_disable_instance_public_access.py
@@ -18,17 +18,17 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsIamDeleteInactiveAccessKeys(RemediationBase):
-    """Remediation that deletes inactive user access keys"""
+class AwsRdsDisableInstancePublicAccess(RemediationBase):
+    """Remediation that disables public access for an RDS instance"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'IAM.DeleteInactiveAccessKeys'
+        return 'RDS.DisableInstancePublicAccess'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
@@ -36,8 +36,4 @@ class AwsIamDeleteInactiveAccessKeys(RemediationBase):
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        client = session.client('iam')
-        if 'AccessKey1Active' in resource['CredentialReport'] and not resource['CredentialReport']['AccessKey1Active']:
-            client.delete_access_key(UserName=resource['UserName'], AccessKeyId=resource['CredentialReport']['AccessKey1Id'])
-        if 'AccessKey2Active' in resource['CredentialReport'] and not resource['CredentialReport']['AccessKey2Active']:
-            client.delete_access_key(UserName=resource['UserName'], AccessKeyId=resource['CredentialReport']['AccessKey2Id'])
+        session.client('rds').modify_db_instance(DBInstanceIdentifier=resource['Id'], PubliclyAccessible=False)
