+++ internal/compliance/remediation_aws/src/app/remediations/aws_rds_disable_snapshot_public_access.py
@@ -18,32 +18,24 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsEc2EnableVpcFlowLogsToS3(RemediationBase):
-    """Remediation that enables VPC Flow logs to S3 bucket"""
+class AwsRdsDisableSnapshotPublicAccess(RemediationBase):
+    """Remediation that disables public access for RDS instance snapshot"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'EC2.EnableVpcFlowLogsToS3'
+        return 'RDS.DisableSnapshotPublicAccess'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
-        return {'TargetBucketName': '', 'TargetPrefix': '', 'TrafficType': 'ALL'}
+        return {}
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        response = session.client('ec2').create_flow_logs(
-            ResourceIds=[
-                resource['Id'],
-            ],
-            ResourceType='VPC',
-            TrafficType=parameters['TrafficType'],
-            LogDestinationType='s3',
-            LogDestination='arn:aws:s3:::{}/{}'.format(parameters['TargetBucketName'], parameters['TargetPrefix'])
-        )
-        if 'Unsuccessful' in response:
-            raise Exception(response['Unsuccessful'][0])
+        client = session.client('rds')
+        for snapshot_attrs in resource['SnapshotAttributes']:
+            client.modify_db_snapshot_attribute(DBSnapshotIdentifier=snapshot_attrs['Id'], AttributeName='restore', ValuesToRemove=['all'])
