+++ internal/compliance/remediation_aws/src/app/remediations/aws_rds_enable_auto_minor_version_upgrade.py
@@ -18,22 +18,26 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsS3EnableBucketVersioning(RemediationBase):
-    """Remediation that enables versioning for an S3 bucket"""
+class AwsRdsEnableAutoMinorVersionUpgrade(RemediationBase):
+    """Remediation that enables Auto Minor Version upgrade for RDS instances"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'S3.EnableBucketVersioning'
+        return 'RDS.EnableAutoMinorVersionUpgrade'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
-        return {}
+        return {'ApplyImmediately': 'true'}
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        session.client('s3').put_bucket_versioning(Bucket=resource['Name'], VersioningConfiguration={'Status': 'Enabled'})
+        session.client('rds').modify_db_instance(
+            DBInstanceIdentifier=resource['Id'],
+            AutoMinorVersionUpgrade=True,
+            ApplyImmediately=parameters['ApplyImmediately'].lower() == 'true'
+        )
