+++ internal/compliance/remediation_aws/src/app/remediations/aws_s3_enable_bucket_logging.py
@@ -18,22 +18,30 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsDdbEncryptTable(RemediationBase):
-    """Remediation that creates a KMS key and uses it to encrypt DDB table"""
+class AwsS3EnableBucketLogging(RemediationBase):
+    """Remediation that enables S3 bucket access logging"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'DDB.EncryptTable'
+        return 'S3.EnableBucketLogging'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
-        return {}
+        return {'TargetBucket': '', 'TargetPrefix': ''}
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        session.client('dynamodb').update_table(TableName=resource['Name'], SSESpecification={'Enabled': True, 'SSEType': 'KMS'})
+        session.client('s3').put_bucket_logging(
+            Bucket=resource['Name'],
+            BucketLoggingStatus={
+                'LoggingEnabled': {
+                    'TargetBucket': parameters['TargetBucket'],
+                    'TargetPrefix': parameters['TargetPrefix']
+                }
+            }
+        )
