+++ internal/compliance/remediation_aws/src/app/remediations/aws_s3_enable_bucket_versioning.py
@@ -18,17 +18,17 @@ from typing import Any, Dict
 
 from boto3 import Session
 
-from ..app import Remediation
-from ..app.remediation_base import RemediationBase
+from .remediation import Remediation
+from .remediation_base import RemediationBase
 
 
 @Remediation
-class AwsRdsDisableSnapshotPublicAccess(RemediationBase):
-    """Remediation that disables public access for RDS instance snapshot"""
+class AwsS3EnableBucketVersioning(RemediationBase):
+    """Remediation that enables versioning for an S3 bucket"""
 
     @classmethod
     def _id(cls) -> str:
-        return 'RDS.DisableSnapshotPublicAccess'
+        return 'S3.EnableBucketVersioning'
 
     @classmethod
     def _parameters(cls) -> Dict[str, str]:
@@ -36,6 +36,4 @@ class AwsRdsDisableSnapshotPublicAccess(RemediationBase):
 
     @classmethod
     def _fix(cls, session: Session, resource: Dict[str, Any], parameters: Dict[str, str]) -> None:
-        client = session.client('rds')
-        for snapshot_attrs in resource['SnapshotAttributes']:
-            client.modify_db_snapshot_attribute(DBSnapshotIdentifier=snapshot_attrs['Id'], AttributeName='restore', ValuesToRemove=['all'])
+        session.client('s3').put_bucket_versioning(Bucket=resource['Name'], VersioningConfiguration={'Status': 'Enabled'})
